+++ src/BitmapText.js
@@ -1,6 +1,5 @@
 /**
- * A BitmapText object will create a line or multiple lines of text using bitmap font. To
- * split a line you can use '\n', '\r' or '\r\n' in your string. You can generate the fnt files using:
+ * A BitmapText object will create a line or multiple lines of text using bitmap font. To split a line you can use '\n', '\r' or '\r\n' in your string. You can generate the fnt files using:
  * <br>
  * <br>
  * A BitmapText can only be created when the font is loaded
@@ -10,7 +9,6 @@
  * http://www.bmglyph.com/ for mac.
  *
  * @example
- *
  * // in this case the font is in a file called 'desyrel.fnt'
  * let bitmapText = new Tiny.BitmapText("text using a fancy font!", {font: "35px Desyrel", align: "right"});
  *
@@ -20,12 +18,10 @@ class BitmapText extends Tiny.Container {
   /**
    * @param {string} text - The copy that you would like the text to display
    * @param {object} style - The style parameters
-   * @param {string|object} style.font - The font descriptor for the object, can be passed as a string of form
-   *      "24px FontName" or "FontName" or as an object with explicit name/size properties.
+   * @param {string|object} style.font - The font descriptor for the object, can be passed as a string of form "24px FontName" or "FontName" or as an object with explicit name/size properties.
    * @param {string} [style.font.name] - The bitmap font id
    * @param {number} [style.font.size] - The size of the font in pixels, e.g. 24
-   * @param {string} [style.align='left'] - Alignment for multiline text ('left', 'center' or 'right'), does not affect
-   *      single line text
+   * @param {string} [style.align='left'] - Alignment for multiline text ('left', 'center' or 'right'), does not affect single line text
    * @param {number} [style.tint=0xFFFFFF] - The tint color
    */
   constructor(text, style = {}) {
@@ -85,21 +81,29 @@ class BitmapText extends Tiny.Container {
     this._text = text;
 
     /**
-     * The max width of this bitmap text in pixels. If the text provided is longer than the
-     * value provided, line breaks will be automatically inserted in the last whitespace.
+     * The max width of this bitmap text in pixels. If the text provided is longer than the value provided, line breaks will be automatically inserted in the last whitespace.
      * Disable by setting value to 0
      *
      * @member {number}
+     * @private
      */
-    this.maxWidth = 0;
+    this._maxWidth = 0;
 
     /**
-     * The max line height. This is useful when trying to use the total height of the Text,
-     * ie: when trying to vertically align.
+     * The max line height. This is useful when trying to use the total height of the Text, ie: when trying to vertically align.
      *
      * @member {number}
+     * @private
      */
-    this.maxLineHeight = 0;
+    this._maxLineHeight = 0;
+
+    /**
+     * Letter spacing. This is useful for setting the space between characters.
+     *
+     * @member {number}
+     * @private
+     */
+    this._letterSpacing = 0;
 
     /**
      * Text anchor. read-only
@@ -132,42 +136,33 @@ class BitmapText extends Tiny.Container {
     const pos = new Tiny.Point();
     const chars = [];
     const lineWidths = [];
+    const text = this.text.replace(/(?:\r\n|\r)/g, '\n');
+    const textLength = text.length;
+    const maxWidth = this._maxWidth * data.size / this._font.size;
 
     let prevCharCode = null;
     let lastLineWidth = 0;
     let maxLineWidth = 0;
     let line = 0;
-    let lastSpace = -1;
-    let lastSpaceWidth = 0;
+    let lastBreakPos = -1;
+    let lastBreakWidth = 0;
+    let spacesRemoved = 0;
     let maxLineHeight = 0;
 
-    for (let i = 0; i < this.text.length; i++) {
-      const charCode = this.text.charCodeAt(i);
+    for (let i = 0; i < textLength; i++) {
+      const charCode = text.charCodeAt(i);
+      const char = text.charAt(i);
 
-      if (/(\s)/.test(this.text.charAt(i))) {
-        lastSpace = i;
-        lastSpaceWidth = lastLineWidth;
+      if (/(?:\s)/.test(char)) {
+        lastBreakPos = i;
+        lastBreakWidth = lastLineWidth;
       }
 
-      if (/(?:\r\n|\r|\n)/.test(this.text.charAt(i))) {
+      if (char === '\r' || char === '\n') {
         lineWidths.push(lastLineWidth);
         maxLineWidth = Math.max(maxLineWidth, lastLineWidth);
-        line++;
-
-        pos.x = 0;
-        pos.y += data.lineHeight;
-        prevCharCode = null;
-        continue;
-      }
-
-      if (lastSpace !== -1 && this.maxWidth > 0 && pos.x * scale > this.maxWidth) {
-        Tiny.removeItems(chars, lastSpace, i - lastSpace);
-        i = lastSpace;
-        lastSpace = -1;
-
-        lineWidths.push(lastSpaceWidth);
-        maxLineWidth = Math.max(maxLineWidth, lastSpaceWidth);
-        line++;
+        ++line;
+        ++spacesRemoved;
 
         pos.x = 0;
         pos.y += data.lineHeight;
@@ -189,16 +184,39 @@ class BitmapText extends Tiny.Container {
         texture: charData.texture,
         line,
         charCode,
-        position: new Tiny.Point(pos.x + charData.xOffset, pos.y + charData.yOffset),
+        position: new Tiny.Point(pos.x + charData.xOffset + (this._letterSpacing / 2), pos.y + charData.yOffset),
       });
-      lastLineWidth = pos.x + (charData.texture.width + charData.xOffset);
-      pos.x += charData.xAdvance;
+      pos.x += charData.xAdvance + this._letterSpacing;
+      lastLineWidth = pos.x;
       maxLineHeight = Math.max(maxLineHeight, (charData.yOffset + charData.texture.height));
       prevCharCode = charCode;
+
+      if (lastBreakPos !== -1 && maxWidth > 0 && pos.x > maxWidth) {
+        ++spacesRemoved;
+        Tiny.removeItems(chars, 1 + lastBreakPos - spacesRemoved, 1 + i - lastBreakPos);
+        i = lastBreakPos;
+        lastBreakPos = -1;
+
+        lineWidths.push(lastBreakWidth);
+        maxLineWidth = Math.max(maxLineWidth, lastBreakWidth);
+        line++;
+
+        pos.x = 0;
+        pos.y += data.lineHeight;
+        prevCharCode = null;
+      }
     }
 
-    lineWidths.push(lastLineWidth);
-    maxLineWidth = Math.max(maxLineWidth, lastLineWidth);
+    const lastChar = text.charAt(text.length - 1);
+
+    if (lastChar !== '\r' && lastChar !== '\n') {
+      if (/(?:\s)/.test(lastChar)) {
+        lastLineWidth = lastBreakWidth;
+      }
+
+      lineWidths.push(lastLineWidth);
+      maxLineWidth = Math.max(maxLineWidth, lastLineWidth);
+    }
 
     const lineAlignOffsets = [];
 
@@ -252,7 +270,7 @@ class BitmapText extends Tiny.Container {
         this._glyphs[i].y -= this._textHeight * this.anchor.y;
       }
     }
-    this.maxLineHeight = maxLineHeight * scale;
+    this._maxLineHeight = maxLineHeight * scale;
   }
 
   /**
@@ -296,7 +314,6 @@ class BitmapText extends Tiny.Container {
   get tint() {
     return this._font.tint;
   }
-
   set tint(value) {
     this._font.tint = (typeof value === 'number' && value >= 0) ? value : 0xFFFFFF;
 
@@ -312,7 +329,6 @@ class BitmapText extends Tiny.Container {
   get align() {
     return this._font.align;
   }
-
   set align(value) {
     this._font.align = value || 'left';
 
@@ -330,7 +346,6 @@ class BitmapText extends Tiny.Container {
   get anchor() {
     return this._anchor;
   }
-
   set anchor(value) {
     if (typeof value === 'number') {
       this._anchor.set(value);
@@ -347,7 +362,6 @@ class BitmapText extends Tiny.Container {
   get font() {
     return this._font;
   }
-
   set font(value) {
     if (!value) {
       return;
@@ -374,7 +388,6 @@ class BitmapText extends Tiny.Container {
   get text() {
     return this._text;
   }
-
   set text(value) {
     value = value.toString() || ' ';
     if (this._text === value) {
@@ -385,8 +398,36 @@ class BitmapText extends Tiny.Container {
   }
 
   /**
-   * The width of the overall text, different from fontSize,
-   * which is defined in the style object
+   * The max width of this bitmap text in pixels. If the text provided is longer than the value provided, line breaks will be automatically inserted in the last whitespace.
+   * Disable by setting value to 0
+   *
+   * @member {number}
+   */
+  get maxWidth() {
+    return this._maxWidth;
+  }
+  set maxWidth(value) {
+    if (this._maxWidth === value) {
+      return;
+    }
+    this._maxWidth = value;
+    this.dirty = true;
+  }
+
+  /**
+   * The max line height. This is useful when trying to use the total height of the Text, ie: when trying to vertically align.
+   *
+   * @member {number}
+   * @readonly
+   */
+  get maxLineHeight() {
+    this.validate();
+
+    return this._maxLineHeight;
+  }
+
+  /**
+   * The width of the overall text, different from fontSize, which is defined in the style object
    *
    * @member {number}
    * @readonly
@@ -398,8 +439,22 @@ class BitmapText extends Tiny.Container {
   }
 
   /**
-   * The height of the overall text, different from fontSize,
-   * which is defined in the style object
+   * Additional space between characters.
+   *
+   * @member {number}
+   */
+  get letterSpacing() {
+    return this._letterSpacing;
+  }
+  set letterSpacing(value) {
+    if (this._letterSpacing !== value) {
+      this._letterSpacing = value;
+      this.dirty = true;
+    }
+  }
+
+  /**
+   * The height of the overall text, different from fontSize, which is defined in the style object
    *
    * @member {number}
    * @readonly
@@ -415,19 +470,35 @@ class BitmapText extends Tiny.Container {
    *
    * @static
    * @param {XMLDocument} xml - The XML document data.
-   * @param {Tiny.Texture} texture - Texture with all symbols.
+   * @param {Object.<string, Tiny.Texture>|Tiny.Texture|Tiny.Texture[]} textures - List of textures for each page. If providing an object, the key is the `<page>` element's `file` attribute in the FNT file.
    * @return {Object} Result font object with font, size, lineHeight and char fields.
    */
-  static registerFont(xml, texture) {
+  static registerFont(xml, textures) {
     const data = {};
     const info = xml.getElementsByTagName('info')[0];
     const common = xml.getElementsByTagName('common')[0];
-    const res = texture.baseTexture.resolution || Tiny.settings.RESOLUTION;
+    const pages = xml.getElementsByTagName('page');
+    const res = Tiny.getResolutionOfUrl(pages[0].getAttribute('file'), Tiny.settings.RESOLUTION);
+    const pagesTextures = {};
 
     data.font = info.getAttribute('face');
     data.size = parseInt(info.getAttribute('size'), 10);
     data.lineHeight = parseInt(common.getAttribute('lineHeight'), 10) / res;
     data.chars = {};
+
+    // Single texture, convert to list
+    if (textures instanceof Tiny.Texture) {
+      textures = [textures];
+    }
+
+    // Convert the input Texture, Textures or object
+    // into a page Texture lookup by "id"
+    for (let i = 0; i < pages.length; i++) {
+      const id = pages[i].getAttribute('id');
+      const file = pages[i].getAttribute('file');
+
+      pagesTextures[id] = textures instanceof Array ? textures[i] : textures[file];
+    }
 
     // parse letters
     const letters = xml.getElementsByTagName('char');
@@ -435,10 +506,10 @@ class BitmapText extends Tiny.Container {
     for (let i = 0; i < letters.length; i++) {
       const letter = letters[i];
       const charCode = parseInt(letter.getAttribute('id'), 10);
-
+      const page = letter.getAttribute('page') || 0;
       const textureRect = new Tiny.Rectangle(
-        (parseInt(letter.getAttribute('x'), 10) / res) + (texture.frame.x / res),
-        (parseInt(letter.getAttribute('y'), 10) / res) + (texture.frame.y / res),
+        (parseInt(letter.getAttribute('x'), 10) / res) + (pagesTextures[page].frame.x / res),
+        (parseInt(letter.getAttribute('y'), 10) / res) + (pagesTextures[page].frame.y / res),
         parseInt(letter.getAttribute('width'), 10) / res,
         parseInt(letter.getAttribute('height'), 10) / res
       );
@@ -448,8 +519,8 @@ class BitmapText extends Tiny.Container {
         yOffset: parseInt(letter.getAttribute('yoffset'), 10) / res,
         xAdvance: parseInt(letter.getAttribute('xadvance'), 10) / res,
         kerning: {},
-        texture: new Tiny.Texture(texture.baseTexture, textureRect),
-
+        texture: new Tiny.Texture(pagesTextures[page].baseTexture, textureRect),
+        page,
       };
     }